refactor(api): Extract metrics_types.rs (PrometheusMetrics)
- Create metrics_types.rs with PrometheusMetrics type and IntoResponse impl
- Add 2 unit tests for metrics creation and response
- metrics.rs: 181 -> 158 lines (-13% reduction)
- All handler DTOs now have consistent *_types.rs pattern
- Total API lib tests: 341
