+++ edgequake/crates/edgequake-api/src/handlers/mod.rs
@@ -19,6 +19,7 @@ pub mod health_types;
 pub mod lineage;
 pub mod lineage_types;
 pub mod metrics;
+pub mod metrics_types;
 pub mod ollama;
 pub mod ollama_types;
 pub mod pipeline;
@@ -53,6 +54,7 @@ pub use health_types::*;
 pub use lineage::*;
 pub use lineage_types::*;
 pub use metrics::*;
+pub use metrics_types::*;
 pub use ollama::*;
 pub use ollama_types::*;
 pub use pipeline::*;
